Moved LiteralRewriter
